+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Looking for RedminPortal for Laravel 4.2? Visit the [v0.1 Branch](https://github
 11. Rearrange menu to split Physical product store and Digital product store (issue #52).
 12. Promotions, Products and Portfolios now allow translation (issue #9).
 13. Use Laravel authentication, removed Sentry dependency (issue #81).
+14. Allow multiple Category images (issue #79).
 
 ### Bug fixes:
 1. Translation to support any language (issue #83).